+++ python/python/glide/async_commands/core.py
@@ -740,6 +740,39 @@ class CoreCommands(Protocol):
             ),
         )
 
+    async def lindex(
+        self,
+        key: str,
+        index: int,
+    ) -> Optional[str]:
+        """
+        Returns the element at `index` in the list stored at `key`.
+
+        The index is zero-based, so 0 means the first element, 1 the second element and so on.
+        Negative indices can be used to designate elements starting at the tail of the list.
+        Here, -1 means the last element, -2 means the penultimate and so forth.
+
+        See https://redis.io/commands/lindex/ for more details.
+
+        Args:
+            key (str): The key of the list.
+            index (int): The index of the element in the list to retrieve.
+
+        Returns:
+            Optional[str]: The element at `index` in the list stored at `key`.
+                If `index` is out of range or if `key` does not exist, None is returned.
+
+        Examples:
+            >>> await client.lindex("my_list", 0)
+                'value1'  # Returns the first element in the list stored at 'my_list'.
+            >>> await client.lindex("my_list", -1)
+                'value3'  # Returns the last element in the list stored at 'my_list'.
+        """
+        return cast(
+            Optional[str],
+            await self._execute_command(RequestType.Lindex, [key, str(index)]),
+        )
+
     async def rpush(self, key: str, elements: List[str]) -> int:
         """
         Inserts all the specified values at the tail of the list stored at `key`.